requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg[binary]==3.2.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
